+++ example/d2048/game.dart
@@ -4,6 +4,7 @@ import 'dart:html';
 import 'dart:math';
 import 'package:boarding/grid.dart';
 import 'package:boarding/boarding.dart';
+import 'package:boarding/util.dart';
 
 part 'model/grid.dart';
 part 'view/board.dart';