+++ example/dash/game.dart
@@ -1,9 +1,9 @@
 library dash;
 
 import 'dart:html';
-import 'package:boarding/grid.dart' show Direction;
 import 'package:boarding/pieces.dart';
 import 'package:boarding/boarding.dart';
+import 'package:boarding/util.dart' show Direction;
 
 part 'view/board.dart';