+++ example/invaders/game.dart
@@ -1,9 +1,9 @@
-library drop;
+library invaders;
 
 import 'dart:html';
 import 'package:boarding/pieces.dart';
 import 'package:boarding/boarding.dart';
-import 'package:boarding/util.dart' show Direction;
+import 'package:boarding/util.dart';
 
 part 'view/board.dart';